+++ main/src/main/java/com/bloatit/model/managers/MemberManager.java
@@ -42,7 +42,7 @@ public final class MemberManager {
 
     /**
      * Gets the member by login.
-     * 
+     *
      * @param login the login
      * @return the member or <code>null</code> if not found.
      */
@@ -65,7 +65,7 @@ public final class MemberManager {
 
     /**
      * Tells if a Login exists.
-     * 
+     *
      * @param login the login
      * @return <code>true</code>, if it exists, <code>false</code> otherwise.
      */
@@ -75,7 +75,7 @@ public final class MemberManager {
 
     /**
      * Tells if Email exists.
-     * 
+     *
      * @param email the email
      * @return <code>true</code>, if it exists, <code>false</code> otherwise.
      */
@@ -85,7 +85,7 @@ public final class MemberManager {
 
     /**
      * Gets the member by id.
-     * 
+     *
      * @param id the id
      * @return the member or <code>null</code> if not found.
      */
@@ -95,7 +95,7 @@ public final class MemberManager {
 
     /**
      * Gets all the members.
-     * 
+     *
      * @return the all the members in the DB.
      */
     public static PageIterable<Member> getAll() {
@@ -104,7 +104,7 @@ public final class MemberManager {
 
     /**
      * Gets all the members ordered by name.
-     * 
+     *
      * @return the all the members in the DB.
      */
     public static PageIterable<Member> getAllMembersOrderByName() {
@@ -120,10 +120,10 @@ public final class MemberManager {
 
     /**
      * Gets the number of members.
-     * 
+     *
      * @return the members count
      */
-    public static int getMembersCount() {
+    public static Long getMembersCount() {
         return DBRequests.count(DaoMember.class);
     }
